+++ src/redturtle/volto/restapi/deserializer/blocks.py
@@ -9,7 +9,7 @@ from zope.component import adapter
 from zope.interface import implementer
 
 
-EXCLUDE_KEYS = ["@type", "token", "value", "@id"]
+EXCLUDE_KEYS = ["@type", "token", "value", "@id", "query"]
 EXCLUDE_TYPES = ["title", "listing", "calendar", "searchEvents"]
 
 
@@ -32,7 +32,7 @@ class GenericResolveUIDDeserializer(object):
 
     def fix_urls_in_block(self, block):
         if isinstance(block, str):
-            return path2uid(context=self.context, link=block)
+            return self.get_uid_from_path(link=block)
         if block.get("@type", "") in EXCLUDE_TYPES:
             return block
         if "UID" in block.keys():
@@ -44,7 +44,7 @@ class GenericResolveUIDDeserializer(object):
             if key in EXCLUDE_KEYS:
                 continue
             if isinstance(val, str):
-                block[key] = path2uid(context=self.context, link=val)
+                block[key] = self.get_uid_from_path(link=val)
             elif isinstance(val, list):
                 block[key] = [self.fix_urls_in_block(x) for x in val]
             elif isinstance(val, dict):
@@ -52,11 +52,42 @@ class GenericResolveUIDDeserializer(object):
                     entity_map = val.get("entityMap", {})
                     for entity_map in entity_map.values():
                         url = entity_map["data"].get("url", "").strip("/")
-                        entity_map["data"]["url"] = path2uid(
-                            context=self.context, link=url
-                        )
+                        entity_map["data"]["url"] = self.get_uid_from_path(link=url)
                 else:
                     block[key] = self.fix_urls_in_block(block=val)
+        return block
+
+    def get_uid_from_path(self, link):
+        """get_uid_from_path.
+
+        :param link:
+        """
+        try:
+            return path2uid(context=self.context, link=link)
+        except IndexError:
+            # the value (link) is not a valid path
+            return link
+
+
+class TableResolveUIDDeserializer(object):
+    """ """
+
+    order = 210  # after standard ones
+    block_type = "table"
+
+    def __init__(self, context, request):
+        self.context = context
+        self.request = request
+
+    def __call__(self, block):
+        for row in block.get("table", {}).get("rows", []):
+            for cell in row.get("cells", []):
+                for entity in cell.get("value", {}).get("entityMap", {}).values():
+                    if entity.get("type") == "LINK":
+                        href = entity.get("data", {}).get("url", "")
+                        entity["data"]["url"] = path2uid(
+                            context=self.context, link=href
+                        )
         return block
 
 
@@ -69,4 +100,16 @@ class GenericResolveUIDDeserializerContents(GenericResolveUIDDeserializer):
 @implementer(IBlockFieldDeserializationTransformer)
 @adapter(IPloneSiteRoot, IRedturtleVoltoLayer)
 class GenericResolveUIDDeserializerRoot(GenericResolveUIDDeserializer):
+    """Deserializer for site-root"""
+
+
+@implementer(IBlockFieldDeserializationTransformer)
+@adapter(IBlocks, IRedturtleVoltoLayer)
+class TableResolveUIDDeserializerContents(TableResolveUIDDeserializer):
+    """Deserializer for content-types that implements IBlocks behavior"""
+
+
+@implementer(IBlockFieldDeserializationTransformer)
+@adapter(IPloneSiteRoot, IRedturtleVoltoLayer)
+class TableResolveUIDDeserializerRoot(TableResolveUIDDeserializer):
     """Deserializer for site-root"""